refactor(client): create config with common constants
